feat:主机信息新增字段填充

+++ src/main/java/xpu/ctrl/docker/vo/HostEntityVO.java
@@ -2,26 +2,28 @@ package xpu.ctrl.docker.vo;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class HostEntityVO {
     private String hostName;
 
     private String hostIp;
 
-    private Long logical_cores;
-    private Long physical_cores;
+    private BigInteger logical_cores;
+    private BigInteger physical_cores;
 
-    private Long disk_total;
-    private Long disk_free;
+    private BigInteger disk_total;
+    private BigInteger disk_free;
 
-    private Long bootTime;
-    private Long upTime;
+    private BigInteger bootTime;
+    private BigInteger upTime;
 
     private String hostOs;
     private String hostPlatformOs;
     private String hostKernelVersion;
-    private Long memTotal;
-    private Long memFree;
+    private BigInteger memTotal;
+    private BigInteger memFree;
 
     private String hostStatusStr;
